split up classes in basic data types into vector3.h, transform.h, collision-vector.h and basic-data-types includes these
