+++ include/math/transform.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "math/vector3.h"
+
+namespace physics
+{
+
+    struct Transform
+    {
+        vector3f pos;
+        float rotation;
+    };
+    
+}
